# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * NyrethaStore API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface StoreInfo {
  storeName: string;
  description: string;
  serverIp: string;
  discordUrl: string;
  currency: string;
  /** @nullable */
  logoUrl?: string | null;
}

export interface ServerStatus {
  online: boolean;
  playerCount: number;
  maxPlayers: number;
  /** @nullable */
  version?: string | null;
}

export interface Package {
  id: number;
  name: string;
  description: string;
  price: number;
  currency: string;
  category: string;
  categorySlug: string;
  /** @nullable */
  imageUrl?: string | null;
  featured: boolean;
  /** @nullable */
  salePricePercent?: number | null;
}

export interface Category {
  id: number;
  name: string;
  slug: string;
  /** @nullable */
  description?: string | null;
  packageCount: number;
}

export interface CartItem {
  packageId: number;
  packageName: string;
  price: number;
  currency: string;
  quantity: number;
  /** @nullable */
  imageUrl?: string | null;
}

export interface CartItemInput {
  packageId: number;
  quantity: number;
}

export interface Cart {
  items: CartItem[];
  total: number;
  currency: string;
  itemCount: number;
}

export interface RecentPayment {
  id: number;
  username: string;
  packageName: string;
  amount: number;
  currency: string;
  createdAt: string;
  /** @nullable */
  avatarUrl?: string | null;
}

export interface TopDonator {
  username: string;
  totalAmount: number;
  currency: string;
  /** @nullable */
  avatarUrl?: string | null;
}

export interface GiftcardCheckInput {
  code: string;
}

export interface GiftcardBalance {
  code: string;
  balance: number;
  currency: string;
  valid: boolean;
}

export type ListPackagesParams = {
category?: string;
};

export type GetRecentPaymentsParams = {
limit?: number;
};