# lib/api-zod/src/generated/types/cart.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * NyrethaStore API
 * OpenAPI spec version: 0.1.0
 */
import type { CartItem } from './cartItem';

export interface Cart {
  items: CartItem[];
  total: number;
  currency: string;
  itemCount: number;
}