# lib/api-zod/src/generated/types/cartItem.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * NyrethaStore API
 * OpenAPI spec version: 0.1.0
 */

export interface CartItem {
  packageId: number;
  packageName: string;
  price: number;
  currency: string;
  quantity: number;
  /** @nullable */
  imageUrl?: string | null;
}