# lib/api-zod/src/generated/types/giftcardBalance.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * NyrethaStore API
 * OpenAPI spec version: 0.1.0
 */

export interface GiftcardBalance {
  code: string;
  balance: number;
  currency: string;
  valid: boolean;
}