# lib/api-zod/src/generated/types/package.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * NyrethaStore API
 * OpenAPI spec version: 0.1.0
 */

export interface Package {
  id: number;
  name: string;
  description: string;
  price: number;
  currency: string;
  category: string;
  categorySlug: string;
  /** @nullable */
  imageUrl?: string | null;
  featured: boolean;
  /** @nullable */
  salePricePercent?: number | null;
}