# lib/api-zod/src/generated/types/recentPayment.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * NyrethaStore API
 * OpenAPI spec version: 0.1.0
 */

export interface RecentPayment {
  id: number;
  username: string;
  packageName: string;
  amount: number;
  currency: string;
  createdAt: string;
  /** @nullable */
  avatarUrl?: string | null;
}