# lib/api-zod/src/generated/types/serverStatus.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * NyrethaStore API
 * OpenAPI spec version: 0.1.0
 */

export interface ServerStatus {
  online: boolean;
  playerCount: number;
  maxPlayers: number;
  /** @nullable */
  version?: string | null;
}