# lib/api-zod/src/generated/types/storeInfo.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * NyrethaStore API
 * OpenAPI spec version: 0.1.0
 */

export interface StoreInfo {
  storeName: string;
  description: string;
  serverIp: string;
  discordUrl: string;
  currency: string;
  /** @nullable */
  logoUrl?: string | null;
}