# lib/api-zod/src/generated/types/topDonator.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * NyrethaStore API
 * OpenAPI spec version: 0.1.0
 */

export interface TopDonator {
  username: string;
  totalAmount: number;
  currency: string;
  /** @nullable */
  avatarUrl?: string | null;
}